requirements.txt
alembic==1.11.1
anyio==3.7.0
asyncio==3.4.3
asyncpg==0.27.0
certifi==2023.5.7
click==8.1.3
databases==0.7.0
dnspython==2.3.0
email-validator==2.0.0.post2
exceptiongroup==1.1.1
fastapi==0.97.0
greenlet==2.0.2
h11==0.14.0
httpcore==0.17.2
httpx==0.24.1
idna==3.4
iniconfig==2.0.0
Mako==1.2.4
MarkupSafe==2.1.3
packaging==23.1
pluggy==1.0.0
psycopg2==2.9.6
pydantic==1.10.9
PyJWT==2.7.0
pytest==7.3.2
python-dotenv==1.0.0
sniffio==1.3.0
SQLAlchemy==1.4.48
starlette==0.27.0
StrEnum==0.4.10
tomli==2.0.1
typing_extensions==4.6.3
uvicorn==0.22.0
